clean up join-queue page
- created HeaderWBackBtn component
- put form content into Form component
- update styling according to new design
