--- src/components/EnterHelp/EnterHelp.module.css
@@ -1,53 +1,0 @@
-.selection {
-  margin: 1rem 0;
-  font-family: "Poppins";
-  font-style: normal;
-  font-weight: 500;
-  font-size: 14px;
-  line-height: 21px;
-  color: #464646;
-  width: 100%;
-}
-
-.required {
-  color: #008CFF;
-}
-
-.inputBtn {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-}
-
-.unClicked {
-  font-family: "Poppins";
-  font-style: normal;
-  font-weight: 500;
-  font-size: 14px;
-  line-height: 21px;
-  padding: 15px;
-  margin: 10px;
-  background: #E5EBF9;
-  border-radius: 20px;
-  border: none;
-  color: #4B4B4B;
-  cursor: pointer;
-}
-.unClicked input {
-  position: absolute;
-  opacity: 0;
-}
-
-.Clicked {
-  background: #008CFF;
-  border: 1px solid #FFFFFF;
-  border-radius: 20px;
-  color: white;
-  padding: 15px;
-  margin: 10px;
-  cursor: pointer;
-}
-.Clicked input {
-  position: absolute;
-  opacity: 0;
-}/*# sourceMappingURL=EnterHelp.module.css.map */
