+++ src/components/HeaderWBackBtn/HeaderWBackBtn.module.css
@@ -58,37 +58,34 @@ body {
   vertical-align: baseline;
 }
 
-.container {
-  background: #FCFCFF;
-  font-family: inherit;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-  margin: 0;
-  min-height: 100vh;
-  background: #F5F8FF;
-}
-
-.circle {
-  width: 50px;
-  height: 50px;
-  background: pink;
-  margin-left: 1rem;
-  opacity: 0;
-}
-
-.card {
-  background: rgba(255, 255, 255, 0.9);
-  border: 2px solid #FFFFFF;
-  border-radius: 30px;
+.header {
+  margin-bottom: 20px;
+  margin-top: 1rem;
   padding: 1rem 2rem;
+  display: flex;
+  justify-content: center;
+  display: flex;
+  width: 100%;
+}
+.header h5 {
+  font-family: "Poppins";
+  font-style: normal;
+  font-weight: 500;
+  font-size: 15px;
+  line-height: 121.52%;
+  color: #3A3B3D;
+}
+
+.headerContainer {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
   width: 50%;
-  margin-bottom: 3rem;
+  justify-content: space-between;
 }
 
 @media (max-width: 768px) {
-  .card {
-    width: 75%;
+  .headerContainer {
+    width: 90%;
   }
-}/*# sourceMappingURL=joinQueue.module.css.map */
+}/*# sourceMappingURL=HeaderWBackBtn.module.css.map */
