+++ src/components/HeaderWBackBtn/HeaderWBackBtn.module.scss
@@ -1,0 +1,38 @@
+@use "../../App";
+
+body {
+    @include App.body;
+}
+
+.header {
+    margin-bottom: 20px;
+    margin-top: 1rem;
+    padding: 1rem 2rem;
+    display: flex;
+    justify-content: center;
+    display: flex;
+    width: 100%;
+
+    h5 {
+        font-family: 'Poppins';
+        font-style: normal;
+        font-weight: 500;
+        font-size: 15px;
+        line-height: 121.52%;
+        color: App.$darkGray;
+    }
+}
+
+.headerContainer {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    width: 50%;
+    justify-content: space-between;
+}
+
+@media (max-width: 768px) {
+    .headerContainer {
+        width: 90%;
+    }
+}
