+++ src/components/courses/ CoursesList.tsx
@@ -46,24 +46,26 @@ export const CoursesList = () => {
 
 
   return (
-    <section className={ style.coursesListContainer}>
+    <section className={style.coursesListContainer}>
       <Toolbar disableGutters>
-        <h5>Host OH </h5> 
+        <h5>Host OH </h5>
         &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-        <h5 style={{color: "#008CFF"}}>TA</h5>
+        <h5 style={{ color: "#008CFF" }}>TA</h5>
       </Toolbar>
-      
-      <div className = { style.coursesContainer }>{renderedInstructorCourses}</div>
-      <br/>
+
+      <div className={style.coursesContainer}>{renderedInstructorCourses}</div>
+      <br />
       <Toolbar disableGutters>
-        <h5>Join Queue </h5> 
+        <a href="/user/student/comp40/check-queue">
+          <h5>Join Queue </h5>
+        </a>
         &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-        <h5 style={{color: "#008CFF"}}>Student</h5>
+        <h5 style={{ color: "#008CFF" }}>Student</h5>
       </Toolbar>
-      
-      <div className = { style.coursesContainer }>{renderedStudentCourses}</div>
+
+      <div className={style.coursesContainer}>{renderedStudentCourses}</div>
     </section>
-  )
+  );
 }
