+++ src/components/title/Title.module.css
@@ -11,6 +11,7 @@
   font-weight: 600;
   font-size: 17px;
   line-height: 130%;
+  cursor: pointer;
   background: transparent;
   border: none;
   color: #008CFF;
