+++ src/components/title/Title.module.scss
@@ -12,6 +12,7 @@
   font-weight: 600;
   font-size: 17px;
   line-height: 130%;
+  cursor: pointer;
 
   background: transparent;
   // text-decoration: none;
