+++ src/views/JoinQueue/joinQueue.module.scss
@@ -4,42 +4,9 @@ body {
     @include App.body;
 }
 
-.form {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.joinQueue {
-    margin: 1rem 14rem;
-}
-
-.header {
-    margin-bottom: 20px;
-    padding: 1rem 2rem;
-    justify-content: center;
-    display: flex;
-
-    background: #FFFFFF;
-    box-shadow: 0px 2px 7px rgba(191, 191, 191, 0.27);
-
-    h5 {
-        font-family: 'Poppins';
-        font-style: normal;
-        font-weight: 500;
-        font-size: 14px;
-        line-height: 21px;
-        text-align: center;
-        color: #3A3B3D;
-    }
-}
-
-.headerContainer {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    width: 70%;
-    justify-content: space-between;
+.container {
+    @include App.container;
+    background: #F5F8FF;
 }
 
 .circle {
@@ -50,131 +17,17 @@ body {
     opacity: 0;
 }
 
-.input {
-    background: #FFFFFF;
-    border: 0.9px solid #C2C2C2;
-    border-radius: 7px;
-    color: #5b5b5b;
-    font-family: 'Poppins';
-    font-style: normal;
-    font-weight: 500;
-    font-size: 14px;
-    line-height: 21px;
-    padding: 1rem;
-    width: 95%;
-    margin: 10px 0;
-}
-
-.select {
-    background: #FFFFFF;
-    border: 0.9px solid #C2C2C2;
-    border-radius: 7px;
-    color: #5b5b5b;
-    font-family: 'Poppins';
-    font-style: normal;
-    font-weight: 500;
-    font-size: 14px;
-    line-height: 21px;
-    padding: 0.5rem;
-    margin: 10px 0;
-    min-width: 10rem;
-    text-align: center;
-}
-
-.submitBtn {
-    background: #606783;
+.card {
+    background: rgba(255, 255, 255, 0.9);
     border: 2px solid #FFFFFF;
     border-radius: 30px;
-    padding: 1.3rem 4rem;
-    color: white;
-    margin: 0 2rem 5rem 2rem;
-    cursor: pointer;
-}
-
-#form:valid .input {
-    background: green;
+    padding: 1rem 2rem;
+    width: 50%;
+    margin-bottom: 3rem;
 }
 
 @media (max-width: 768px) {
-    .joinQueue{
-        margin: 1rem 3rem;
-    }
-    .select {
-        width: 100%;
-    }
-    .headerContainer {
-        width: 100%;
+    .card {
+        width: 75%;
     }
 }
-
-::placeholder {
-    /* Chrome, Firefox, Opera, Safari 10.1+ */
-    color: #C2C2C2;
-    opacity: 1;
-    /* Firefox */
-}
-
-:-ms-input-placeholder {
-    /* Internet Explorer 10-11 */
-    color: #C2C2C2;
-}
-
-
-.selection {
-    margin: 1rem 0;
-    font-family: 'Poppins';
-    font-style: normal;
-    font-weight: 500;
-    font-size: 14px;
-    line-height: 21px;
-    color: #464646;
-    width: 100%;
-}
-
-.required {
-    color: #008CFF;
-}
-
-.inputBtn {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-}
-
-.unClicked {
-    font-family: 'Poppins';
-    font-style: normal;
-    font-weight: 500;
-    font-size: 14px;
-    line-height: 21px;
-
-    padding: 15px;
-    margin: 10px;
-
-    background: #E5EBF9;
-    border-radius: 20px;
-    border: none;
-    color: #4B4B4B;
-    cursor: pointer;
-
-    input {
-        position: absolute;
-        opacity: 0;
-    }
-}
-
-.Clicked {
-    background: #008CFF;
-    border: 1px solid #FFFFFF;
-    border-radius: 20px;
-    color: white;
-    padding: 15px;
-    margin: 10px;
-    cursor: pointer;
-
-    input {
-        position: absolute;
-        opacity: 0;
-    }
-}
-
