+++ src/views/enroll/Enroll.tsx
@@ -1,17 +1,16 @@
 import styles from "./enroll.module.css";
-import { BackBtn } from "../../components/BackBtn/BackBtn";
+import Tooltip from "@mui/material/Tooltip";
+import { HeaderWBackBtn } from "../../components/HeaderWBackBtn/HeaderWBackBtn"
 
 export function Enroll() {
-  const path = "/user/courses";
+
   return (
     <div className={styles.container}>
-      <div className={styles.header}>
-        <div className={styles.headerContainer}>
-          <BackBtn path={path} />
-          <h5>Add Class</h5>
-          <div className={styles.circle}></div>
-        </div>
-      </div>
+      <HeaderWBackBtn
+        path={"/user/courses"}
+        title={"Add Class"}
+        color={{ color: "white" }}
+      />
       <div className={styles.card}>
         <div className={styles.Header1}>Enter Class Code to Add a Class</div>
         <input
@@ -21,10 +20,14 @@ export function Enroll() {
           name="classCode"
           placeholder="👉 Enter Code"
         />
-        <div className={styles.Subtitle}>Code can be found in your syllabus</div>
-        <a className={styles.btn} href="/user/courses">
-          Add Class
-        </a>
+        <div className={styles.Subtitle}>
+          Code can be found in your syllabus
+        </div>
+        <Tooltip title="Add Class">
+          <a className={styles.btn} href="/user/courses">
+            Add Class
+          </a>
+        </Tooltip>
       </div>
     </div>
   );
